\name{lung2}
\alias{lung2}
\docType{data}
\title{Four compartment microarray data generate by Martin Zand lab,
  2012a.  These are the new ones generated by }
\description{
  \code{lung2} is the lung2 data generated by Martin Zand lab.  Shuang did
  the significance test (the second version) and selected 2,751
  significant genes.  The first time point (day zero minus) is removed.
}
\usage{data(lung2)}
\format{
  This is a 2,751 by .. matrix.
}
\details{
  See technical report for more details.
}

\references{

  Qiu, X., Wu, H., and Wu, S. (2012). A New Information Criterion Based
on Mixture Langevin Distribution for Clustering Circular Data with
Application to Time Course Genomic Data.  \emph{Journal of American
Statistical Association} (submitted). Temporary URL:
\url{http://www.rochester.edu/}.\cr

}

\author{
  Xing Qiu, Hulin Wu, Shuang Wu.
}

\examples{
library(DelayActivation)
data(lung2); Time <- 0:10
head(lung2)
}
\keyword{datasets}
