\name{PostProc}
\alias{PostProc}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
  A function to process results generated by \code{DelayActivation}. 
}

\description{ \code{PostProc} ...  }

\usage{ PostProc(firstrun.info, manual.delay=NULL, manual.regular=NULL,
                     act.min=30, modes.min=50, cluster.min=10)
}
%- maybe also 'usage' for other objects documented here.

\arguments{
  \item{firstrun.info}{The results produced by \code{FirstRun}.}

  \item{manual.delay}{A list of the index of genes to be set to "Delay"
    manually. The actual code is: \code{Delay[manual.delay] <- "Delay"}.}
  \item{manual.regular}{A list of the index of genes to be set to "Regular"
    manually. The actual code is: \code{Delay[manual.regular] <- "Regular"}.}

  \item{act.min}{Minimum number of genes in each activation day.  If
  number of genes with a particular activation day is less or equal to
  \option{act.min}, this day will be merged to a near day. Please
  consult the attached report file for more details. Default to 30.}

  \item{modes.min}{Minimum number of genes in each mode class.  If
  number of genes with a particular mode is less or equal to
  \option{modes.min}, this mode class will be merged to a nearby mode.
  Please consult the attached report file for more details. Default to 50.}

  \item{cluster.min}{Minimum number of genes in each cluster.  If number
  of genes in a particular cluster is less or equal to \option{act.min},
  this cluster will be merged to a nearby cluster.  Please consult the
  attached report file for more details. Default to 10.}  
}

\details{ \code{PostProc} reads the clusters generated by
  \code{FirstRun}, then merge small clusters (clusters with less than
  \option{cluster.min} genes; days of activation with less than
  \option{act.min} genes; modes with less than \option{modes.min}
  genes. See the attached report for more details) with most similar
  large clusters. The The actual algorithm is a bit complicated so
  please consult the attached report for more details.

  You can also specify manual overrides of delay/regular genes here.
}

\value{ A list of three components
  \item{fittded.curves}{A functional data (fd) object of smoothed gene
    expression (as functions of time) produced by function
    \code{smooth.basis()} from package \pkg{fda}.}

  \item{clust.info}{A data frame containing the results obtained from
  \code{\link{DelayDetect}}, \code{\link{ModesDetect}},
  \code{\link{ActivationDetect}}}.  An additional column,
  \option{Clust}, is provided for convenience. \option{Clust} is defined
  as \code{paste(Delay, UpDown, ActDay, Modes, sep=".")}, the unique 
  identifier for clusters defined by these four criteria.
}

\references{some references.}

\author{
  Xing Qiu, Hulin Wu, Shuang Wu.
}

\seealso{\code{\link{DelayDetect}}, \code{\link{ModesDetect}},
  \code{\link{ActivationDetect}}, \code{\link{FirstRun}},
  \code{\link{plot.cluster}} \code{\link{plotpdf.clusters}}. }

\examples{
library(DelayActivation)
data(lung2); Time <- 0:10

mybasis <- create.bspline.basis(range(Time), length(Time)+4-2, 4, Time)
smoothing.par <- fdPar(mybasis, Lfdobj=2, lambda=5.0)

first.results <- FirstRun(lung2, Time, smoothing.par)
rr2 <- PostProc(first.results, cluster.min=60)

\dontrun{
plotpdf.clusters(lung2, Time, rr2, "firstrun.pdf",
                xlim=c(0,10), ylim=c(-3,3), 
                xlab="Time (day)", ylab="Expression")
}
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{clustering}
